% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hepatitisA}
\alias{hepatitisA}
\title{Hepatitis A data}
\format{A data frame with 83 rows and  three variables:
\describe{
    \item{t}{ Age of the individual}
    \item{freq1}{ Number of individuals of age t that are seropositive for Hepatitis A}
    \item{freq2}{ Total number of individuals of age t}
}}
\usage{
hepatitisA
}
\description{
A dataset on the prevalence of hepatitis A in individuals from Bulgaria
with age ranging from 1 to 86 years.
The data consists of a cross-sectional survey conducted in 1964.
}
\references{
Keiding, N. (1991). Age-specic incidence and prevalence: a statistical perspective. J. Roy. Statist. Soc. Ser. A,154(3):371-412.
}
\keyword{datasets}
