% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rubella}
\alias{rubella}
\title{Rubella data}
\format{A data frame with 225 rows and  three variables:
\describe{
    \item{t}{ Age of the individual at the time of testing for immunization}
    \item{freq1}{ Number of individuals of age t that are immune for Rubella}
    \item{freq2}{ Total number of individuals of age t}
}}
\usage{
rubella
}
\description{
A dataset on the prevalence of rubella in 230 Austrian males
older than three months for whom the exact date of birth was known.
Each individual was tested at the Institute of Virology, Vienna
during the period 1–25 March 1988 for immunization against Rubella.
}
\references{
Keiding, N., Begtrup, K., Scheike, T., and Hasibeder, G. (1996). Estimation from current status data in continuous time. Lifetime Data Anal., 2:119-129.
}
\keyword{datasets}
