% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeConfIntervals}
\alias{ComputeConfIntervals}
\title{Confidence Intervals}
\usage{
ComputeConfIntervals(input, x, alpha)
}
\arguments{
\item{input}{DataFrame with three variables:
\describe{
    \item{xx}{Observation points t sorted in ascending order}
    \item{freq1}{Frequency of observation t satisfying \eqn{y \le t}}
    \item{freq2}{Frequency of observation t}
}}

\item{x}{numeric vector}

\item{alpha}{level for confidence intervals}
}
\value{
List with 4 variables:
\enumerate{
    \item{MLE}{ maximum likelihood estimator}
    \item{SMLE}{ smoothed maximum likelihood estimator}
    \item{h}{ local bandwidths for points of grid}
    \item{CI}{ confidence intervals on points of grid}
    }
}
