% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeMLE}
\alias{ComputeMLE}
\title{Maximum Likelihood Estimator}
\usage{
ComputeMLE(data)
}
\arguments{
\item{data}{DataFrame with three variables:
\describe{
    \item{t}{Observation points t sorted in ascending order}
    \item{freq1}{Frequency of observation t satisfying \eqn{y \le t}}
    \item{freq2}{Frequency of observation t}
}}
}
\value{
DataFrame with two variables :
\describe{
    \item{x}{jumplocations of the MLE}
    \item{mle}{MLE evaluated at the jumplocations}
}
}
