% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeMLE}
\alias{ComputeMLE}
\title{Maximum Likelihood Estimator}
\usage{
ComputeMLE(data)
}
\arguments{
\item{data}{DataFrame with three variables:
\describe{
    \item{t}{Observation points t sorted in ascending order}
    \item{freq1}{Frequency of observation t satisfying \eqn{y \le t}}
    \item{freq2}{Frequency of observation t}
}}
}
\value{
DataFrame with two variables :
\describe{
    \item{x}{jumplocations of the MLE}
    \item{mle}{MLE evaluated at the jumplocations}
}
}
\description{
The function ComputeMLE computes the Maximum likelihood Estimator of the distribution function under current status data.
}
\examples{
library(Rcpp)
library(curstatCI)

# sample size
n <- 1000

# Uniform data  U(0,2)
set.seed(2)
y <- runif(n,0,2)
t <- runif(n,0,2)
delta <- as.numeric(y <= t)

A<-cbind(t[order(t)], delta[order(t)], rep(1,n))
mle <-ComputeMLE(A)
plot(mle$x, mle$mle,type ='s', ylim=c(0,1), main= "",ylab="",xlab="",las=1)

}
\references{
The nonparametric bootstrap for the current status model, Groeneboom, P. and Hendricx, K. Electronical Journal of Statistics (2017)
}
