% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeSMLE}
\alias{ComputeSMLE}
\title{Smoothed Maximum Likelihood Estimator}
\usage{
ComputeSMLE(data, x, h)
}
\arguments{
\item{data}{DataFrame with three variables:
\describe{
    \item{t}{Observation points t sorted in ascending order}
    \item{freq1}{Frequency of observation t satisfying \eqn{y \le t}}
    \item{freq2}{Frequency of observation t}
}}

\item{x}{a number or vector}

\item{h}{bandwidth}
}
\value{
SMLE(x)
}
